requirements.txt
absl-py==2.1.0
Adafruit-Blinka==8.32.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-connectionmanager==1.0.0
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==3.0.0
adafruit-circuitpython-typing==1.10.2
Adafruit-PlatformDetect==3.60.0
Adafruit-PureIO==1.1.11
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
flatbuffers==20181003210633
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
requests==2.31.0
requests-oauthlib==1.3.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sysv-ipc==1.1.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-cpu-aws==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.3.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
